added custom hooks for fetching movies
