+++ src/components/Loading.jsx
@@ -1,0 +1,9 @@
+function Loading() {
+  return (
+    <section>
+      <h2 className="text-center text-3xl py-4 font-semibold">Loading...</h2>
+    </section>
+  );
+}
+
+export default Loading;
